scripts/requirements.txt
feedparser==6.0.11
httpx==0.27.0
beautifulsoup4==4.12.3
trafilatura==1.9.0
python-frontmatter==1.0.1
python-slugify==8.0.4
pillow==10.4.0
ruamel.yaml==0.18.6
jinja2==3.1.4
pydantic==2.8.2
readabilipy==0.2.0
tqdm==4.66.4
regex==2024.5.15
# Added for local AI pipeline scripts
PyYAML==6.0.2
requests==2.32.3
requests-cache==1.2.0
